[FIX] stock: unlink scrap move on discard
Steps to Reproduce:
- Install Inventory with demo data.
- Inventory > Operations > Adjustments > Scrap.
- Create a new scrap with product 'Large Cabinet' and quantity 1000.
- Click Confirm > (In the warning wizard) Click Discard> Click Confirm.

Error:
ValueError - Expected singleton: stock.move()

Before 19.2;
Since https://github.com/odoo/odoo/commit/c361c3778ef4755b4760039a4fd8f9ed88294b64, in the Shop Floor flow (button_scrap), the scrap is created
first, and discarding an insufficient warning will unlink the scrap order.
However, when using the scrap form view (via adjustments), the scrap was
not unlinked on discard.

Later https://github.com/odoo/odoo/commit/1c7d80a10b5d7db1c4163166bf52b3f3c77044ba, the condition was removed during refactoring, causing
the scrap move is to be unlinked in all flows.

Fix:
This commit adds a context key to ensure that the scrap move is only unlinked
in the action_scrap flow.

sentry-7336640883
opw-6128188

+++ addons/stock/wizard/stock_warn_insufficient_qty.py
@@ -46,4 +46,5 @@ class StockWarnInsufficientQtyScrap(models.TransientModel):
         return self.with_context(clean_context(self.env.context)).scrap_move_id._action_scrap()
 
     def action_cancel(self):
-        return self.scrap_move_id.sudo().unlink()
+        if self.env.context.get('unlink_on_discard'):
+            self.scrap_move_id.sudo().unlink()
